Update tsql scripts to use MONEY on the MinimumBid, Price and Value fields

+++ sql/dbo.Auction.sql
@@ -4,7 +4,7 @@
     [EndTime]     DATETIME NOT NULL,
     [ClientId]    INT      NOT NULL,
     [ProductId]   INT      NOT NULL,
-    [MinimumBid]  INT      NOT NULL,
+    [MinimumBid]  MONEY    NOT NULL,
     [IsCompleted] BIT      DEFAULT ((0)) NOT NULL,
     PRIMARY KEY CLUSTERED ([AuctionId] ASC)
 );
