+++ sql/dbo.Bid.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Bid] (
     [BidId]     INT      NOT NULL,
-    [Value]     INT      NOT NULL,
+    [Value]     MONEY    NOT NULL,
     [Time]      DATETIME NOT NULL,
     [ClientId]  INT      NOT NULL,
     [AuctionId] INT      NOT NULL,
