+++ sql/dbo.Product.sql
@@ -2,7 +2,7 @@
     [ProductId]   INT             NOT NULL,
     [Name]        NVARCHAR (64)   NOT NULL,
     [Description] NVARCHAR (2048) NOT NULL,
-    [Price]       INT             NOT NULL,
+    [Price]       MONEY           NOT NULL,
     [AuctionId]   INT             NOT NULL,
     [ClientId]    INT             DEFAULT ((0)) NOT NULL,
     [Images]      NVARCHAR (2048) NULL,
